Refactor imports in CallClient and WebSocket modules for improved organization and clarity. Update index exports to include CONNECTION_QUALITY in call constants. Ensure consistent import paths across components utilizing these modules.
