+++ webapp/src/clients/call/index.ts
@@ -2,6 +2,6 @@
 // See LICENSE.txt for license information.
 
 export {default} from './call_client';
-export {CALL_EVENT} from './constants';
+export {CALL_EVENT, CONNECTION_QUALITY} from './constants';
 export type {CallClientStats, ConnectPayload, TrackMetadata} from './types';
 export {DisconnectReason} from 'livekit-client';
