+++ webapp/src/clients/websocket/types.ts
